:hammer: Extract User32 utility functions into focused WindowsProcessUtils, WindowsFocusUtils, WindowsIconUtils objects
