+++ app/src/desktopMain/kotlin/com/crosspaste/app/WinAppWindowManager.kt
@@ -7,7 +7,7 @@ import com.crosspaste.path.UserDataPathProvider
 import com.crosspaste.platform.windows.WinAppInfo
 import com.crosspaste.platform.windows.WinAppInfoCaches
 import com.crosspaste.platform.windows.WindowFocusRecorder
-import com.crosspaste.platform.windows.api.User32
+import com.crosspaste.platform.windows.WindowsFocusUtils
 import com.crosspaste.platform.windows.api.User32.Companion.INSTANCE
 import com.crosspaste.platform.windows.api.WndEnumProc
 import com.sun.jna.Native
@@ -30,7 +30,7 @@ class WinAppWindowManager(
     val mainHWND: HWND?
         get() {
             if (_cachedMainHWND == null) {
-                _cachedMainHWND = User32.findPasteWindow(mainWindowTitle)
+                _cachedMainHWND = WindowsFocusUtils.findPasteWindow(mainWindowTitle)
             }
             return _cachedMainHWND
         }
@@ -38,7 +38,7 @@ class WinAppWindowManager(
     val searchHWND: HWND?
         get() {
             if (_cachedSearchHWND == null) {
-                _cachedSearchHWND = User32.findPasteWindow(searchWindowTitle)
+                _cachedSearchHWND = WindowsFocusUtils.findPasteWindow(searchWindowTitle)
             }
             return _cachedSearchHWND
         }
@@ -46,7 +46,7 @@ class WinAppWindowManager(
     val bubbleHWND: HWND?
         get() {
             if (_cachedBubbleHWND == null) {
-                _cachedBubbleHWND = User32.findPasteWindow(bubbleWindowTitle)
+                _cachedBubbleHWND = WindowsFocusUtils.findPasteWindow(bubbleWindowTitle)
             }
             return _cachedBubbleHWND
         }
@@ -105,7 +105,7 @@ class WinAppWindowManager(
     override suspend fun focusMainWindow(windowTrigger: WindowTrigger) {
         // Wait for the window to be ready, otherwise bringToFront may cause the window to fail to get focus
         delay(500)
-        User32.bringToFront(
+        WindowsFocusUtils.bringToFront(
             windowFocusRecorder.lastWinAppInfo.value?.getThreadId(winAppInfoCaches),
             mainHWND,
         )
@@ -120,7 +120,7 @@ class WinAppWindowManager(
     override suspend fun focusSearchWindow(windowTrigger: WindowTrigger) {
         // Wait for the window to be ready, otherwise bringToFront may cause the window to fail to get focus
         delay(500)
-        User32.bringToFront(
+        WindowsFocusUtils.bringToFront(
             windowFocusRecorder.lastWinAppInfo.value?.getThreadId(winAppInfoCaches),
             searchHWND,
         )
@@ -128,7 +128,7 @@ class WinAppWindowManager(
 
     override suspend fun focusBubbleWindow() {
         delay(500)
-        User32.bringToFront(
+        WindowsFocusUtils.bringToFront(
             windowFocusRecorder.lastWinAppInfo.value?.getThreadId(winAppInfoCaches),
             bubbleHWND,
         )
@@ -158,13 +158,13 @@ class WinAppWindowManager(
                 lazyShortcutKeys.value.shortcutKeysCore.value.keys[PASTE]?.let {
                     it.map { key -> key.rawCode }
                 } ?: listOf()
-            User32.bringToBackAndPaste(
+            WindowsFocusUtils.bringToBackAndPaste(
                 backHWND,
                 windowFocusRecorder.lastWinAppInfo.value?.hwnd,
                 keyCodes,
             )
         } else {
-            User32.backToBack(backHWND, windowFocusRecorder.lastWinAppInfo.value?.hwnd)
+            WindowsFocusUtils.backToBack(backHWND, windowFocusRecorder.lastWinAppInfo.value?.hwnd)
         }
     }
 
@@ -174,7 +174,7 @@ class WinAppWindowManager(
                 it.map { key -> key.rawCode }
             } ?: listOf()
 
-        User32.paste(keyCodes)
+        WindowsFocusUtils.paste(keyCodes)
     }
 
     override fun onMainComposeWindowChanged(window: ComposeWindow?) {
